RandomSpot2 and Sequence2
to allow for alternative definition of ‘random’ spot: picking spot from
spot list. to allow backward compatibility, Spot.RandomSpot()
definitions are replaced with a .createRandomSpot() function, which is
overridden in Sequence2 to call RandomSpot2 instead of RandomSpot

sequence.rand_xy() also created and partitioned out of the
sequence.randomize(). This allows overriding, again, in
Sequence2.rand_xy() which selects a spot from list of random spots.